Add files via upload
Yoo guys, I struggled a bit and tried doing the "login' and "create account" pages for the site. As well as i just linked them correspondingly to the "index" file. Just check it out guys, and let me know what y'all think about it...

I also added the necessary bootstrap folder when uploading, so I hope it'll make the process now easy to see my changes since i used the framework for both pages. There's an unnecessary picture in the login file which is to serve as the logo, but I'll look for a different picture to represent the Cameroon exam system later.... Y'all just *fork* it and edit if need be ehh    : )...